From: Paul Walmsley <pjw@archive.org>
The timeout backoff loop in the PTP2 camera_wait_for_event() function
has a bug that can result in unexpectedly long delays. While polling
the camera for PTP2 events with a timeout of 100 milliseconds,
camera_wait_for_event() could take over 200 milliseconds to complete.
This occurred even when no messages were received from the camera.
This was due to some unusual logic in the timeout backoff loop.

To fix, modify the timeout backoff loop such that it calculates the
time interval between the present time and the last received PTP2
event, and then subtracts that from the current timeout.  This ensures
that 'resttime' decreases as the time interval between the present
time and the last event increases.  Also, modify the code to sleep up
to 20 milliseconds at a time, rather than 20 unconditionally - this
ensures that the code does not wait for longer than it was instructed
to.

This significantly reducess the time needed to capture a frame from
the camera.  This is important for automated applications such as book
digitization, etc.

The development of this patch was sponsored by the Internet Archive.
The
Archive and its partners have digitized several million pages with
gphoto2, using the Canon EOS 5D Mark I and II cameras.  The digitized
texts are available free to the public at:

    http://archive.org/details/texts